+++ Sample/QuickStart/AppDelegate.swift
@@ -45,6 +45,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterD
         // GroupChannel - form type message
         SendbirdUI.config.groupChannel.channel.isFormTypeMessageEnabled = true
         
+//        if #available(iOS 14, *) {
+//            SendbirdUI.config.groupChannel.channel.isMultipleFilesMessageEnabled = true
+//        }
+        
+        
         self.initializeRemoteNotification()
         
         return true