+++ Sources/View/Channel/SBUChatNotificationChannelViewController.swift
@@ -379,7 +379,19 @@ open class SBUChatNotificationChannelViewController: SBUBaseViewController,
             newNotificationInfoView.updateCount(count: self.newNotificationsCount) { [weak self] in
                 guard let self = self else { return }
                 guard let listComponent = self.listComponent else { return }
-                self.chatNotificationChannelModuleDidTapScrollToButton(listComponent, animated: true)
+                
+                let position = SBUGlobals.scrollPostionConfiguration.groupChannel.scrollToNewMessage
+                let options = SBUScrollOptions(
+                    count: self.newNotificationsCount,
+                    position: position,
+                    isInverted: listComponent.tableView.isInverted
+                )
+                
+                self.chatNotificationChannelModule(
+                    listComponent,
+                    didSelectScrollToBottonWithOptions: options,
+                    animated: true
+                )
             }
         }
 
@@ -609,6 +621,22 @@ open class SBUChatNotificationChannelViewController: SBUBaseViewController,
         _ listComponent: SBUChatNotificationChannelModule.List,
         animated: Bool
     ) {
+        let options = SBUScrollOptions(
+            position: .bottom,
+            isInverted: listComponent.tableView.isInverted
+        )
+        self.chatNotificationChannelModule(
+            listComponent,
+            didSelectScrollToBottonWithOptions: options,
+            animated: animated
+        )
+    }
+    
+    func chatNotificationChannelModule(
+        _ listComponent: SBUChatNotificationChannelModule.List,
+        didSelectScrollToBottonWithOptions options: SBUScrollOptions,
+        animated: Bool
+    ) {
         guard self.viewModel?.notifications.isEmpty == false else { return }
         self.newNotificationsCount = 0
         
@@ -625,8 +653,11 @@ open class SBUChatNotificationChannelViewController: SBUBaseViewController,
                 self.viewModel?.reloadNotificationList()
                 self.listComponent?.scrollTableView(to: 0)
             } else {
-                let indexPath = IndexPath(row: 0, section: 0)
-                self.listComponent?.scrollTableView(to: indexPath.row, animated: animated)
+                self.listComponent?.scrollTableView(
+                    to: options.row(),
+                    at: options.at(),
+                    animated: animated
+                )
                 self.updateNewNotificationInfo(hidden: true)
             }
         }